requirements.txt
# Core Dependencies


# Machine Learning
xgboost==2.0.2


# Web Framework
altair==5.2.0

# Model Explainability

# Data Processing

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
pyyaml==6.0.1
joblib==1.3.2
# Core Dependencies
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.15.0

# Feature Store & MLOps (FIXED VERSION)
hopsworks==4.2.9

# Model Explainability
shap==0.44.0

# Visualization & Dashboard
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2

# Utilities
pytz==2023.3
python-dateutil==2.8.2

# Testing
pytest==7.4.3

# Logging
loguru==0.7.2
# Core Dependencies
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.15.0

# Feature Store & MLOps - CORRECTED VERSION
hopsworks==4.2.9

# Model Explainability
shap==0.44.0

# Visualization & Dashboard
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2

# Utilities
pytz==2023.3
python-dateutil==2.8.2

# Testing
pytest==7.4.3

# Logging
loguru==0.7.2